Detect vertical video and don't crop
A simple "hotfix" to prevent the cropping of video that is already vertical. It only detects if the height > width.

Thus it will distort videos with a different aspect ratio than 9:16
In practice vertical videos are almost always in this aspect ratio, so for the time being (until the cropping in general get's improved) this is a simple solution.