Save the video with the largest resolution.
Currently it iterates over all found urls of a video and overwrites the returned value until the last entry. 
This results in low quality video because the order in which the urls are returned is random.

This change ensures only the url of the highest resolution version of a found video is saved.