Error handling for main.py
I added a try-except block around the generate_subtitles function call to catch any exceptions that might occur during subtitle generation. If an exception occurs, it prints an error message and sets subtitles_path to None. I did the same thing for the generate_video function call to catch any exceptions that might occur during the generation of the final video. If an exception occurs, it prints an error message and sets final_video_path to None.